finish ensemble tools, called xBle

+++ PropertyInspectionPrediction/code/model_library.py
@@ -53,7 +53,7 @@ def train_model(path, x_train, y_train, x_test, y_test, feat):
     if model_list.count('dnn') > 0:
         model_type = 'dnn'
         pred_file = "%s/%s_%s.pred.pkl" %(path, feat, model_type)
-        if os.path.exists(pred_file) is False:
+        if config.update_model.count(model_type) > 0 or os.path.exists(pred_file) is False:
             print "%s training..." % model_type
             model_param = config.best_param[model_type]
             model = deep_model()
@@ -68,7 +68,7 @@ def train_model(path, x_train, y_train, x_test, y_test, feat):
     if model_list.count('knn') > 0:
         model_type = 'knn'
         pred_file = "%s/%s_%s.pred.pkl" %(path, feat, model_type)
-        if os.path.exists(pred_file) is False:
+        if config.update_model.count(model_type) > 0 or os.path.exists(pred_file) is False:
             print "%s training..." % model_type
             model_param = config.best_param[model_type]
             n_neighbors = model_param['n_neighbors']
@@ -84,7 +84,7 @@ def train_model(path, x_train, y_train, x_test, y_test, feat):
     if model_list.count('linear') > 0:
         model_type = 'linear'
         pred_file = "%s/%s_%s.pred.pkl" %(path, feat, model_type)
-        if os.path.exists(pred_file) is False:
+        if config.update_model.count(model_type) > 0 or os.path.exists(pred_file) is False:
             print "%s training..." % model_type
             model_param = config.best_param[model_type]
             model = LinearRegression()
@@ -98,7 +98,7 @@ def train_model(path, x_train, y_train, x_test, y_test, feat):
     if model_list.count('logistic') > 0:
         model_type = 'logistic'
         pred_file = "%s/%s_%s.pred.pkl" %(path, feat, model_type)
-        if os.path.exists(pred_file) is False:
+        if config.update_model.count(model_type) > 0 or os.path.exists(pred_file) is False:
             print "%s training..." % model_type
             model_param = config.best_param[model_type]
             model = LogisticRegression()
@@ -112,7 +112,7 @@ def train_model(path, x_train, y_train, x_test, y_test, feat):
     if model_list.count('svr') > 0:
         model_type = 'svr'
         pred_file = "%s/%s_%s.pred.pkl" %(path, feat, model_type)
-        if os.path.exists(pred_file) is False:
+        if config.update_model.count(model_type) > 0 or os.path.exists(pred_file) is False:
             print "%s training..." % model_type
             model_param = config.best_param[model_type]
             model = SVR(C=model_param['C'], epsilon=model_param['epsilon'])
@@ -126,7 +126,7 @@ def train_model(path, x_train, y_train, x_test, y_test, feat):
     if model_list.count('ranksvm') > 0:
         model_type = 'ranksvm'
         pred_file = "%s/%s_%s.pred.pkl" %(path, feat, model_type)
-        if os.path.exists(pred_file) is False:
+        if config.update_model.count(model_type) > 0 or os.path.exists(pred_file) is False:
             print "%s training..." % model_type
             model_param = config.best_param[model_type]
             model = RankSVM().fit( x_train, y_train )
@@ -139,7 +139,7 @@ def train_model(path, x_train, y_train, x_test, y_test, feat):
     if model_list.count('rf') > 0:
         model_type = 'rf'
         pred_file = "%s/%s_%s.pred.pkl" %(path, feat, model_type)
-        if os.path.exists(pred_file) is False:
+        if config.update_model.count(model_type) > 0 or os.path.exists(pred_file) is False:
             print "%s training..." % model_type
             model_param = config.best_param[model_type]
             model = RandomForestRegressor(n_estimators=model_param['n_estimators'])
@@ -153,7 +153,7 @@ def train_model(path, x_train, y_train, x_test, y_test, feat):
     if model_list.count('extratree') > 0:
         model_type = 'extratree'
         pred_file = "%s/%s_%s.pred.pkl" %(path, feat, model_type)
-        if os.path.exists(pred_file) is False:
+        if config.update_model.count(model_type) > 0 or os.path.exists(pred_file) is False:
             print "%s training..." % model_type
             model_param = config.best_param[model_type]
             model = ExtraTreesRegressor(n_estimators=model_param['n_estimators'])
@@ -167,7 +167,7 @@ def train_model(path, x_train, y_train, x_test, y_test, feat):
     if model_list.count('gbf') > 0:
         model_type = 'gbf'
         pred_file = "%s/%s_%s.pred.pkl" %(path, feat, model_type)
-        if os.path.exists(pred_file) is False:
+        if config.update_model.count(model_type) > 0 or os.path.exists(pred_file) is False:
             print "%s training..." % model_type
             model_param = config.best_param[model_type]
             model = GradientBoostingRegressor(n_estimators=model_param['n_estimators'])
@@ -186,7 +186,7 @@ def train_model(path, x_train, y_train, x_test, y_test, feat):
     if model_list.count('xgboost') > 0:
         model_type = 'xgboost'
         pred_file = "%s/%s_%s.pred.pkl" %(path, feat, model_type)
-        if os.path.exists(pred_file) is False:
+        if config.update_model.count(model_type) > 0 or os.path.exists(pred_file) is False:
             print "%s training..." % model_type
             model_param = config.best_param[model_type]
             params = model_param
@@ -194,14 +194,46 @@ def train_model(path, x_train, y_train, x_test, y_test, feat):
             #create a train and validation dmatrices
             xgtrain = xgb.DMatrix(x_train, label=y_train)
             xgval = xgb.DMatrix(x_test, label=y_test)
-            print x_test.shape[0]
-            print len(y_test)
 
             #train using early stopping and predict
             watchlist = [(xgtrain, "train"),(xgval, "val")]
             #model = xgb.train(params, xgtrain, num_rounds, watchlist, early_stopping_rounds=100, feval=gini_metric)
             model = xgb.train(params, xgtrain, num_rounds, watchlist, early_stopping_rounds=100)
             pred_val = model.predict( xgval, ntree_limit=model.best_iteration )
+            with open(pred_file, 'wb') as f:
+                pickle.dump(pred_val, f, -1)
+            print "Done!"
+
+    if model_list.count('xgboost-art') > 0:
+        model_type = 'xgboost-art'
+        pred_file = "%s/%s_%s.pred.pkl" %(path, feat, model_type)
+        if config.update_model.count(model_type) > 0 or os.path.exists(pred_file) is False:
+            print "%s trainning..." % model_type
+            model_param = config.best_param[model_type]
+            params = model_param
+            num_rounds = model_param['num_rounds']
+            offset = int(model_param['valid_size'] * y_train.shape[0])
+            if type(x_train) != np.ndarray:
+                x_train = x_train.toarray()
+                x_test = x_test.toarray()
+            xgtrain = xgb.DMatrix(x_train[offset:, :], label=y_train[offset:])
+            xgval = xgb.DMatrix(x_train[:offset, :], label=y_train[:offset])
+            watchlist = [(xgtrain, "train"), (xgval, "val")]
+            model = xgb.train(params, xgtrain, num_rounds, watchlist, early_stopping_rounds = model_param['early_stopping_rounds'])
+
+            xgtest = xgb.DMatrix(x_test)
+            pred_val1 = model.predict(xgtest, ntree_limit=model.best_iteration)
+
+            # reverse train, and log label
+            x_train_tmp = x_train[::-1, :]
+            y_train_tmp = np.log(y_train[::-1])
+            xgtrain = xgb.DMatrix(x_train_tmp[offset:, :], label=y_train_tmp[offset:])
+            xgval = xgb.DMatrix(x_train_tmp[:offset, :], label=y_train_tmp[:offset])
+            watchlist = [(xgtrain, "train"), (xgval, "val")]
+            model = xgb.train(params, xgtrain, num_rounds, watchlist, early_stopping_rounds = model_param['early_stopping_rounds'])
+
+            pred_val2 = model.predict(xgtest, ntree_limit=model.best_iteration)
+            pred_val = pred_val1*1.5 + pred_val2*8.5
             with open(pred_file, 'wb') as f:
                 pickle.dump(pred_val, f, -1)
             print "Done!"
@@ -215,6 +247,7 @@ def one_model():
     for iter in range(config.kiter):
         for fold in range(config.kfold):
             for feat in feat_names:
+                print "Gen pred for (iter%d, fold%d, %s) cross validation" %(iter, fold, feat)
                 with open("%s/iter%d/fold%d/train.%s.feat.pkl" %(config.data_folder, iter, fold, feat), 'rb') as f:
                     [x_train, y_train] = pickle.load(f)
                 with open("%s/iter%d/fold%d/valid.%s.feat.pkl" %(config.data_folder, iter, fold, feat), 'rb') as f:
@@ -224,6 +257,7 @@ def one_model():
 
     # load feat, train/test
     for feat in feat_names:
+        print "Gen pred for (%s) all test data" %(feat)
         with open("%s/all/train.%s.feat.pkl" %(config.data_folder, feat), 'rb') as f:
             [x_train, y_train] = pickle.load(f)
         with open("%s/all/test.%s.feat.pkl" %(config.data_folder, feat), 'rb') as f:
@@ -361,7 +395,28 @@ def hyperopt_library(model_type, model_param, x_train, y_train, x_test):
         pred_val = model.predict( xgval, ntree_limit=model.best_iteration )
 
     if model_type.count('xgboost-art') > 0:
+        print "%s trainning..." % model_type
+        params = model_param
+        num_rounds = model_param['num_rounds']
+        offset = int(model_param['valid_size'] * y_train.shape[0])
+        xgtrain = xgb.DMatrix(x_train[offset:, :], label=y_train[offset:])
+        xgval = xgb.DMatrix(x_train[:offset, :], label=y_train[:offset])
+        watchlist = [(xgtrain, "train"), (xgval, "val")]
+        model = xgb.train(params, xgtrain, num_rounds, watchlist, early_stopping_rounds = 100)
 
+        xgtest = xgb.DMatrix(x_test)
+        pred_val1 = model.predict(xgtest, ntree_limit=model.best_iteration)
+
+        # reverse train, and log label
+        x_train_tmp = x_train[::-1, :]
+        y_train_tmp = np.log(y_train[::-1])
+        xgtrain = xgb.DMatrix(x_train_tmp[offset:, :], label=y_train_tmp[offset:])
+        xgval = xgb.DMatrix(x_train_tmp[:offset, :], label=y_train_tmp[:offset])
+        watchlist = [(xgtrain, "train"), (xgval, "val")]
+        model = xgb.train(params, xgtrain, num_rounds, watchlist, early_stopping_rounds = 100)
+
+        pred_val2 = model.predict(xgtest, ntree_limit=model.best_iteration)
+        pred_val = pred_val1*1.5 + pred_val2*8.5
 
     return pred_val
 
@@ -384,11 +439,14 @@ if __name__ == '__main__':
     # write your code here
     # apply different model on different feature, generate model library
 
-    ## generate pred by best params
-    #one_model()
+    flag = sys.argv[1]
+    if flag == "train":
+        ## generate pred by best params
+        one_model()
 
-    ## hyper parameter search
-    hyperopt_main()
+    if flag == "hyperopt":
+        ## hyper parameter search
+        hyperopt_main()
